Add ability to use last played mission and...
Fixed a bug where rerandomizing the battlefield would cause the sectors
to cap faster.
Fixed the weather settings not working.
Fixed a bug where an admin could teleport players to their starting
zones again by pressing the end safestart option. There's now a check in
place to prevent this.
Changed date in mission.sqm to one with a full moon.