Move gear removal to loadout_set
In preparation for CBA 2.40 version. After that comes out I will update
the loadouts to add gear after a one frame delay. From what I remember
there's an arma bug that causes item desync when uniforms are removed
and added back on the same frame. It waits until CBA 2.4 since they are
adding an execOnNextFrame function with that update.